DB/Schemas/requirements.txt
python-dotenv
psycopg2
psycopg2-binary


